Proof of concept for image manipulation using OpenCV.  This program will
provide a blue tint over the image.  To remove this effect simply remove
the word 'blue' from line 41.  Another observation is that when not
manipulating the image the time between start and end seems to increase
at a steady rate. Yet when manipulating an image the time seems to stay
the same between 60-80 milliseconds per frame